HACK: css2600: Don't use unsigned long
BZ: 174635

The size of unsigned long is dependent on the environment. Make sure we
always use 32 bits on both host and SP.

<URL:https://icggerrit.ir.intel.com/#/c/2464/>

+++ drivers/media/pci/css2600/lib2401/bxt_sandbox/isyspoc/interface/ia_css_isysapi_fw_types.h
@@ -337,7 +337,7 @@ struct ia_css_isys_frame_buff_set {
 struct ia_css_isys_resp_info {
 	enum ia_css_isys_resp_type type;
 	int error;
-	unsigned long timestamp[2];
+	uint32_t timestamp[2];
 	unsigned int stream_handle;
 	struct ia_css_isys_output_pin pin;
 };
